Crudle upgrade

+++ app/modules/setup/controllers/DataModelController.php
@@ -1,16 +1,18 @@
 <?php
 
-namespace crudle\setup\controllers;
+namespace crudle\app\setup\controllers;
 
-use crudle\main\controllers\base\BaseCrudController;
+use crudle\app\main\controllers\action\EditRow;
+use crudle\app\main\controllers\action\Index;
+use crudle\app\main\controllers\base\BaseCrudController;
 use Yii;
-use yii\base\Model;
 use yii\data\ActiveDataProvider;
 use yii\web\NotFoundHttpException;
-use crudle\setup\models\DataModel;
-use crudle\setup\models\DataModelSearch;
-use crudle\setup\models\DataModelField;
-use crudle\setup\models\DataModelFieldSearch;
+use crudle\app\setup\models\DataModel;
+use crudle\app\setup\models\search\DataModelSearch;
+use crudle\app\setup\models\DataModelField;
+use crudle\app\setup\models\search\DataModelFieldSearch;
+use yii\helpers\ArrayHelper;
 
 /**
  * DataModelController implements the CRUD actions for DataModel model.
@@ -29,6 +31,14 @@ class DataModelController extends BaseCrudController
         return DataModelSearch::class;
     }
 
+    public function actions()
+    {
+        return [
+            'index' => Index::class,
+            'edit-row' => EditRow::class,
+        ];
+    }
+
     /**
      * Creates a new DataModel model.
      * If creation is successful, the browser will be redirected to the 'read' view.
@@ -37,7 +47,7 @@ class DataModelController extends BaseCrudController
     public function actionCreate($id = null)
     {
         $this->model = new DataModel();
-        $this->detailModels = [];
+        $modelDetails = $this->getDetailModels();
 
         $formDetails = Yii::$app->request->post('DataModelField', []);
         foreach ($formDetails as $i => $formDetail) 
@@ -45,46 +55,54 @@ class DataModelController extends BaseCrudController
             $modelDetail = new DataModelField(['scenario' => DataModelField::SCENARIO_BATCH_ACTION]);
             $modelDetail->attributes = $formDetail;
             $modelDetail->actionType = DataModelField::ACTION_TYPE_CREATE;
-            $this->detailModels[] = $modelDetail;
+            $modelDetails[] = $modelDetail;
         }
 
         $this->fieldDataProvider = new ActiveDataProvider([
-            'query' => DataModelField::find()->where(['data_model' => '']),
+            'query' => DataModelField::find()->where(['model_name' => '']),
         ]);
 
-        $this->fieldDataProvider->setModels( !empty($this->detailModels) ? $this->detailModels :  [new DataModelField(['scenario' => DataModelField::SCENARIO_BATCH_ACTION])] );
+        $this->fieldDataProvider->setModels( !empty($modelDetails) ? $modelDetails['dataModelFields'] :  [new DataModelField(['scenario' => DataModelField::SCENARIO_BATCH_ACTION])] );
 
         // handling if the addRow button has been pressed
         if ( isset( Yii::$app->request->post()['addRow']) ) 
         {
-            $this->model->load(Yii::$app->request->post());
-            $this->detailModels[] = new DataModelField(['scenario' => DataModelField::SCENARIO_BATCH_ACTION]);
-            $this->fieldDataProvider->setModels( $this->detailModels );
+            $this->model->load(Yii::$app->request->post(), 'DataModel');
+            unset($modelDetails['dataModelFields']); // To-Do: fix later by refactoring this action method
+            $modelDetail = new DataModelField(['scenario' => DataModelField::SCENARIO_BATCH_ACTION]);
+            $modelDetail->model_name = $this->model->id;
+            $modelDetails[] = $modelDetail;
+            $this->fieldDataProvider->setModels( $modelDetails );
 
-            return $this->render('@app_main/views/_crud/index', [
+            return $this->render('@appMain/views/crud/index', [
                 'model' => $this->model,
             ]);
         }
 
-        if ($this->model->load(Yii::$app->request->post())) 
+        if ($this->model->load(Yii::$app->request->post(), 'DataModel'))
         {
-            if (Model::validateMultiple($this->detailModels) && $this->model->validate()) 
+            $hasDetailModels = $this->loadDetailModels(DataModelField::class, $formDetails);
+            if ($this->validateDetailModels() && $this->model->validate()) 
             {
-                // use AR/DB transaction here ?
-                if ( $this->model->save(false) ) 
+                if ( $this->model->save(false) && $hasDetailModels) 
                 {
-                    foreach($this->detailModels as $modelDetail) 
+                    foreach($this->detailModels['DataModelField'] as $modelDetail) 
                     {
-                        $modelDetail->data_model = $this->model->name;
+                        $modelDetail->model_name = $this->model->id;
                         $modelDetail->save(false);
                     }
-                    $this->model->createTable();
+                    if ((bool) $this->model->is_table)
+                        $this->model->createTable();
                 }
-                return $this->redirect(['update', 'id' => $this->model->name]);
+                return $this->redirect(['update', 'id' => $this->model->id]);
+            }
+            else
+            {
+                Yii::$app->session->setFlash( 'errors', $this->validationErrors);
             }
         }
 
-        return $this->render('@app_main/views/_crud/index', [
+        return $this->render('@appMain/views/crud/index', [
             'model' => $this->model,
         ]);
     }
@@ -99,15 +117,18 @@ class DataModelController extends BaseCrudController
     public function actionUpdate($id)
     {
         $model = $this->findModel($id);
-        $modelDetails = $model->dataModelFields;
+        $modelDetails = $this->getDetailModels();
 
         $formDetails = Yii::$app->request->post('DataModelField', []);
         foreach ($formDetails as $i => $formDetail) 
         {
             //loading the models if they are not new
-            if (isset($formDetail['name']) && isset($formDetail['actionType']) && $formDetail['actionType'] != DataModelField::ACTION_TYPE_CREATE) {
+            if (isset($formDetail['field_name']) &&
+                isset($formDetail['actionType']) && $formDetail['actionType'] != DataModelField::ACTION_TYPE_CREATE)
+            {
                 //making sure that it is actually a child of the main model
-                $modelDetail = DataModelField::findOne(['name' => $formDetail['name'], 'data_model' => $model->name]);
+                $modelDetail = DataModelField::findOne(['field_name' => $formDetail['field_name'], 
+                                                        'model_name' => $model->id]);
                 $modelDetail->setScenario(DataModelField::SCENARIO_BATCH_ACTION);
                 $modelDetail->setAttributes($formDetail);
                 $modelDetails[$i] = $modelDetail;
@@ -122,55 +143,89 @@ class DataModelController extends BaseCrudController
         }
 
         $this->fieldDataProvider = new ActiveDataProvider([
-            'query' => DataModelField::find()->where(['data_model' => '']),
+            'query' => DataModelField::find()->where(['model_name' => '']),
         ]);
 
         //handling if the addRow button has been pressed
         if ( isset( Yii::$app->request->post()['addRow']) ) 
         {
-            $model->load(Yii::$app->request->post());
-            $modelDetails[] = new DataModelField(['scenario' => DataModelField::SCENARIO_BATCH_ACTION]);
+            $model->load(Yii::$app->request->post(), 'DataModel');
+            unset($modelDetails['dataModelFields']); // To-Do: fix later by refactoring this action method
+            $modelDetail = new DataModelField(['scenario' => DataModelField::SCENARIO_BATCH_ACTION]);
+            $modelDetail->model_name = $model->id;
+            $modelDetails[] = $modelDetail;
             $this->fieldDataProvider->setModels( $modelDetails );
 
             $this->model = $model;
 
-            return $this->render('@app_main/views/_crud/index', [
+            return $this->render('@appMain/views/crud/index', [
                 'model' => $model,
             ]);
         }
 
-        if ($model->load(Yii::$app->request->post())) 
-        {
-            if (Model::validateMultiple($modelDetails) && $model->validate()) 
+        $postData = Yii::$app->request->post();
+        if ($postData) {
+            if ($model->load($postData, 'DataModel'))
             {
-                if ( $model->save(false) ) 
+                $detailModels = $this->_reIndexFieldList($formDetails);
+                $hasDetailModels = $this->loadDetailModels(DataModelField::class, $detailModels);
+                if ($this->validateDetailModels() && $model->validate()) 
                 {
-                    foreach($modelDetails as $modelDetail) 
+                    if ( $model->save(false) && $hasDetailModels) 
                     {
-                        //details that has been flagged for deletion will be deleted
-                        if ($modelDetail->actionType == DataModelField::ACTION_TYPE_DELETE) {
-                            $modelDetail->delete();
-                        } else {
-                            //new or updated records go here
-                            $modelDetail->data_model = $model->name;
-                            $modelDetail->save(false);
-                        }                        
+                        foreach($this->detailModels['DataModelField'] as $modelDetail) 
+                        {
+                            // details that has been flagged for deletion will be deleted
+                            if ($modelDetail->actionType == DataModelField::ACTION_TYPE_DELETE) {
+                                $modelDetail->delete();
+                            } else {
+                                // new or updated records go here
+                                $modelDetail->model_name = $model->id;
+                                $modelDetail->save(false);
+                            }
+                        }
+                        // $model->updateTable(); // TODO: define method in model
                     }
-                    // $model->updateTable(); // TODO: define method in model
+                    return $this->redirect(['update', 'id' => $model->id]);
                 }
-                return $this->redirect(['update', 'id' => $model->name]);
-            }
-            else
-            {
-                Yii::$app->session->setFlash( 'error', $model->errors);
+                else
+                {
+                    Yii::$app->session->setFlash( 'errors', $this->validationErrors);
+                }
             }
         }
 
-        $this->fieldDataProvider->setModels( !empty($modelDetails) ? $modelDetails :  [new DataModelField(['scenario' => DataModelField::SCENARIO_BATCH_ACTION])] );
+        $this->fieldDataProvider->setModels( !empty($modelDetails) ? $modelDetails['dataModelFields'] :  [new DataModelField(['scenario' => DataModelField::SCENARIO_BATCH_ACTION])] );
 
         $this->model = $model;
 
-        return $this->render('@app_main/views/_crud/index', [
+        return $this->render('@appMain/views/crud/index', [
+            'model' => $model,
+        ]);
+    }
+
+    /**
+     * Reads an existing DataModel model.
+     * @param string $id
+     * @return mixed
+     * @throws NotFoundHttpException if the model cannot be found
+     */
+    public function actionRead($id)
+    {
+        $model = $this->findModel($id);
+        $modelDetails = $this->getDetailModels();
+
+        $this->fieldDataProvider = new ActiveDataProvider([
+            'query' => DataModelField::find()->where(['model_name' => $id]),
+        ]);
+
+        $this->fieldDataProvider->setModels(
+            !empty($modelDetails['dataModelFields']) ?
+                $modelDetails['dataModelFields'] :
+                [new DataModelField(['scenario' => DataModelField::SCENARIO_BATCH_ACTION])] );
+        $this->model = $model;
+
+        return $this->render('@appMain/views/crud/index', [
             'model' => $model,
         ]);
     }
@@ -183,5 +238,53 @@ class DataModelController extends BaseCrudController
         return [
             'fieldSearchModel' => $searchModel,
         ];
+    }
+
+    private function _reIndexFieldList(&$formDetails)
+    {
+        $colIndex = 0;
+        foreach ($formDetails as $i => $formDetail) 
+        {
+            $colIndex = $colIndex + 1;
+            $formDetail['col_index'] = $colIndex;
+            $formDetails[$i] = $formDetail;
+        }
+
+        return $formDetails;
+    }
+
+    public function actionReIndexFieldList()
+    {
+        if ( Yii::$app->request->isAjax )
+        {
+            $modelId = Yii::$app->request->post('modelId');
+            $model = DataModel::findOne($modelId);
+            if (is_null($model))
+                $model = new DataModel();
+            $rows = Yii::$app->request->post('gridData');
+            $modelDetails = []; $colIndex = 0;
+            foreach ($rows as $row) {
+                $modelDetail = DataModelField::findOne($row['id']);
+                if (is_null($modelDetail))
+                    $modelDetail = new DataModelField();
+                $modelDetail->col_index = $colIndex + 1;
+                $modelDetails[] = $modelDetail;
+            }
+
+            $fieldDataProvider = new ActiveDataProvider([
+                'query' => DataModelField::find()->where(['model_name' => '']),
+            ]);
+            $fieldDataProvider->setModels(!empty($modelDetails) ?
+                    $modelDetails :
+                    [new DataModelField(['scenario' => DataModelField::SCENARIO_BATCH_ACTION])] );
+
+            return 
+                $this->controller->renderAjax('@appSetup/views/data_model/field/list_columns', [
+                    'dataProvider' => $fieldDataProvider,
+                    'model' => $model,
+                ]);
+        }
+        // else
+        Yii::$app->end();
     }
 }
